Split monolithic test files into per-module files mirroring source structure
Replaces test_chess.py and test_coverage.py (and the broken pandas-based
test_positions.py) with one test file per source module:

  test_board.py         — Board utilities + en passant, castling, promotion,
                          has_moved, legal_moves_safe
  test_positions.py     — PieceMoves (all piece types), _slide capture_only,
                          _jump checkers-style captures
  test_win_conditions.py — GameResult, ChessWinCondition, CheckersWinCondition
  test_svg_renderer.py  — _parse_color, _parse_path, render_svg
  test_game.py          — Graphics coord/draw helpers, Game.end_turn

All 167 tests pass.
